Fix paths to the workflow modules
