+++ cg/cli/workflow/__init__.py
@@ -1,7 +1,1 @@
-"""Adds cg cli workflow command and sub commands"""
-
-from cg.cli.workflow.workflow import workflow
-from cg.cli.workflow.balsamic import balsamic
-from cg.cli.workflow.microsalt import microsalt
-from cg.cli.workflow.mip_dna import mip_dna
-from cg.cli.workflow.mip_rna import mip_rna
+"""Module init"""
