+++ cg/cli/workflow/balsamic/__init__.py
@@ -1,0 +1,1 @@
+"""Init for module"""
