+++ cg/cli/workflow/mip_rna/__init__.py
@@ -1,0 +1,1 @@
+"""Init for module"""
