+++ tests/cli/workflow/mip_dna/test_cli_mip_dna.py
@@ -2,7 +2,7 @@
 from datetime import datetime
 
 import pytest
-from cg.cli.store.mip_dna import _add_new_complete_analysis_record
+from cg.cli.workflow.mip_dna.store import _add_new_complete_analysis_record
 from cg.exc import AnalysisDuplicationError
 from cg.store import Store
